Remove required labels flag.
When initally putting in this logic we were unsure if
there were labels other than instance that could need to be
set to empty. After further thought, it's only the job
and instance labels that are mandatory on the prometheus server
side and as job is already mandatory for the pushgateway
only instance needs special casing.