Implement lifetime feature
In some situations it is very useful if you can submit values to
the pushgateway that disappear after a certain while (if they
are not refreshed).

The lifetime is specified by adding a "Lifetime"-Field to your
HTTP-Header. The value is a string that the "ParseDuration"
(golang-builtin) function accepts as valid format.

Implements #19