Reject pushes with timestamps.
Fixes #109

I once thought that some day we'd solve staleness, and there
would be valid use cases for the pushgateway to expose timestamps.
With the recent work on staleness this has turned out not to be the case.
As every single person we have seen pushing timestamps to the pgw
has been doing so incorrectly (usually misguidedly trying to make
Prometheus do push), let's send a stronger signal that this is not the
right way to do things.

With this in place I think we'll have all the mitigation we're going to
get in terms of discourage users from misusing timestamps, and it'll be
tolerable to add timestamp support to client libraries for the very very
few use cases that need this advanced feature (at current count:
collectd, cloudwatch, graphite and inflxudb exporters).

This should in principle have no impact as a) there is no documentation
saying that this is a good idea or even an idea and b) there is no code
support in our libraries for pushing timestamps.