Stop using prometheus.LabelPairSorter
This commit replaces it by an internal implementation. This _is_ code
replication, but only very little and very trivial code. The upside is
that this removes the only known use of prometheus.LabelPairSorter
outside of prometheus/client_golang. (I plan to un-export
prometheus.LabelPairSorter to reduce the large number of exported
types in client_golang.)

Signed-off-by: beorn7 <beorn@soundcloud.com>