Provide support for --web.external-url
This works in conjunction with the `--web.route-prefix` flag to
provide resources at the provided external URL, with the properly
munged `route-prefix`. If an external URL is provided, but a
`route-prefix` is not, this change takes the external URL's Path
and uses that as the `route-prefix`.

Includes a simple test checking for the presence of a known
external-url in the body of the Status handler.

Signed-off-by: Andrew G10z <agwozdziewycz@salesforce.com>