Also log error messages during gathering
Quite a few of users have run into the problem that they see a scrape
error in the Prometheus UI but don't see the error. (It's not
displayed in the Prometheus UI, you have to go to the /metrics
endopint yourself to see it.) Logging it makes the problem much more obvious, because all those users looked at what the PGW is logging, which was nothing in that case.

Signed-off-by: beorn7 <beorn@grafana.com>