Cut v1.0.0
:tada:

v0.10.0 is out for a few days. No complaints so far. I've been running
it myself at Grafana since Friday. Let's do it! The only change
compared to v0.10 is the removal of the legacy storage conversion code
and an update of dependencies, which only really affects the new
prometheus/client_golang.

Signed-off-by: beorn7 <beorn@grafana.com>