Web UI: Fix deletion of groups with empty label values
If an empty label value ends up in the middle of the URL, it leads to
a double-`/`. The Pushgateway requires a single padding `=` character
to avoid the problem.

With this change, we simply always add the padding `=`. It's optional
and can be used at will even with non-empty label values.

Signed-off-by: beorn7 <beorn@grafana.com>