resources: naive rendering for native histograms (#596)
Classic buckets are rendered as before, if there are any. If there are native buckets, a string version of the native histogram is rendered like Prometheus currently does. This approach also handles "dual use" histograms (both classic and native buckets).

A histogram without any buckets is rendered as such, just with count and sum.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>