Adjust to the prometheus/common new world order (v0.66)
Primarily, we need to use `NewTextParser`. While doing that, I have
also removed all the deprecated references to the package-global
variable in the `model` package. (We are still using package-globel
variables in our own `handler` package. But I guess that is fine for
the time being. We needed to pass around the ValidationScheme
everywhere otherwise.)

Signed-off-by: beorn7 <beorn@grafana.com>